Allowing to add extra code between MODBUS request and response
When MODBUS is executed on software serial, it's difficult to maintain
any other operations, because software serial in fact always polls for
the incoming RX data (except when it sends TX data), making all other
delays and interrupts behavior unstable.

A workaround applied here simply assumes that after a MODBUS request
is received, we have an open window to do other stuff before sending
a response. For this purpose we allow to override the request handlers,
such that they could be prepended by user-defined code.